\name{logpep-package}
\alias{logpep-package}
\alias{logpep}
\docType{package}
\title{
\packageTitle{logpep}
}
\description{
\packageDescription{logpep}
}
\details{

\packageDESCRIPTION{logpep}
\packageIndices{logpep}
Create a log file and add a timestamp. Also you can add a note. Finally, if a timestamp is passed to write_time, it returns the difference with the timestamp of the call
}
\author{
\packageAuthor{logpep}

Maintainer: \packageMaintainer{logpep}
}

\keyword{ package }


